Refactor PurchaseMetaPaymentDetails into its own tsx file
